pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "keeper"
version = "0.1.0"
description = "智能运维助手 - 交互式 CLI 工具"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "pydantic>=2.0",
    "click>=8.0",
    "prompt-toolkit>=3.0",
    "pyyaml>=6.0",
    "psutil>=5.9",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "flake8>=6.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
]

[project.scripts]
keeper = "keeper.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["keeper*"]
